refactor: move SignaturePublisher to a dedicated module
