+++ mithril-signer/src/services/aggregator_client.rs
@@ -16,14 +16,13 @@ use mithril_common::{
     messages::{
         AggregatorFeaturesMessage, EpochSettingsMessage, TryFromMessageAdapter, TryToMessageAdapter,
     },
-    StdError, StdResult, MITHRIL_API_VERSION_HEADER, MITHRIL_SIGNER_VERSION_HEADER,
+    StdError, MITHRIL_API_VERSION_HEADER, MITHRIL_SIGNER_VERSION_HEADER,
 };
 
 use crate::entities::SignerEpochSettings;
 use crate::message_adapters::{
     FromEpochSettingsAdapter, ToRegisterSignatureMessageAdapter, ToRegisterSignerMessageAdapter,
 };
-use crate::services::SignaturePublisher;
 
 const JSON_CONTENT_TYPE: HeaderValue = HeaderValue::from_static("application/json");
 
@@ -170,20 +169,6 @@ pub trait AggregatorClient: Sync + Send {
     async fn retrieve_aggregator_features(
         &self,
     ) -> Result<AggregatorFeaturesMessage, AggregatorClientError>;
-}
-
-#[async_trait]
-impl<T: AggregatorClient> SignaturePublisher for T {
-    async fn publish(
-        &self,
-        signed_entity_type: &SignedEntityType,
-        signatures: &SingleSignatures,
-        protocol_message: &ProtocolMessage,
-    ) -> StdResult<()> {
-        self.register_signatures(signed_entity_type, signatures, protocol_message)
-            .await?;
-        Ok(())
-    }
 }
 
 /// AggregatorHTTPClient is a http client for an aggregator
