+++ mithril-signer/src/services/mod.rs
@@ -14,6 +14,7 @@ mod cardano_transactions;
 mod certifier;
 mod epoch_service;
 mod signable_builder;
+mod signature_publisher;
 mod single_signer;
 mod upkeep_service;
 
@@ -24,5 +25,6 @@ pub use cardano_transactions::*;
 pub use certifier::*;
 pub use epoch_service::*;
 pub use signable_builder::*;
+pub use signature_publisher::*;
 pub use single_signer::*;
 pub use upkeep_service::*;
