+++ mithril-signer/src/services/signature_publisher/mod.rs
@@ -1,0 +1,4 @@
+mod http;
+mod interface;
+
+pub use interface::*;
